fetch success

+++ app/src/main/java/com/technorapper/myfriendsbday/data/model/CurrencyListModel.kt
@@ -1,5 +1,3 @@
 package com.technorapper.myfriendsbday.data.model
 
-import java.util.Currency
-
-data class CurrencyListModel(val currency: String, val rate: Double)
+data class CurrencyListModel(val currency: String, val rate: Float)
